Split is_deleted test to a separate class
