+++ sink-connector-lightweight/src/test/java/com/altinity/clickhouse/debezium/embedded/ddl/parser/TableOperationsIT.java
@@ -189,55 +189,4 @@ public class TableOperationsIT {
 
         }
 
-    @ParameterizedTest
-    @CsvSource({
-            "clickhouse/clickhouse-server:latest",
-            "clickhouse/clickhouse-server:22.3"
-    })
-    @DisplayName("Test that validates create table in CH when MySQL has is_deleted columns")
-    public void testIsDeleted(String clickHouseServerVersion) throws Exception {
-
-        AtomicReference<DebeziumChangeEventCapture> engine = new AtomicReference<>();
-
-        ExecutorService executorService = Executors.newFixedThreadPool(1);
-        executorService.execute(() -> {
-            try {
-
-                engine.set(new DebeziumChangeEventCapture());
-                engine.get().setup(ITCommon.getDebeziumProperties(mySqlContainer, clickHouseContainer), new SourceRecordParserService(),
-                        new MySQLDDLParserService(new ClickHouseSinkConnectorConfig(new HashMap<>())),false);
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
-        });
-
-
-        Thread.sleep(10000);
-        Connection conn = ITCommon.connectToMySQL(mySqlContainer);
-        conn.prepareStatement("create table new_table(col1 varchar(255), col2 int, is_deleted int, _sign int)").execute();
-
-        Thread.sleep(10000);
-
-        conn.prepareStatement("insert into new_table values('test', 1, 22, 1)").execute();
-        conn.close();
-        Thread.sleep(10000);
-
-        BaseDbWriter writer = new BaseDbWriter(clickHouseContainer.getHost(), clickHouseContainer.getFirstMappedPort(),
-                "employees", clickHouseContainer.getUsername(), clickHouseContainer.getPassword(), null);
-        ResultSet rs = writer.executeQueryWithResultSet("select * from new_table");
-        boolean recordFound = false;
-        while(rs.next()) {
-            recordFound = true;
-            Assert.assertTrue(rs.getString("col1").equalsIgnoreCase("test"));
-            Assert.assertTrue(rs.getInt("col2") == 1);
-            Assert.assertTrue(rs.getInt("is_deleted") == 22);
-            Assert.assertTrue(rs.getInt("_sign") == 1);
-        }
-        Assert.assertTrue(recordFound);
-
-        if(engine.get() != null) {
-            engine.get().stop();
-        }
-        executorService.shutdown();
-    }
 }
